requirements.txt
# Python dependencies for OVPFH v2.0 Firestore upload script

firebase-admin>=6.0.0
python-dotenv>=1.0.0